fix: cast byte to int for Math.Min (.NET Framework 4.8 ambiguity)
CS0121: Math.Min(byte,byte) vs Math.Min(int,int) ambiguous in net48.

+++ PitLeaguePlugin/Adapters/F1_25/Udp/SessionHistoryPacket.cs
@@ -57,7 +57,7 @@ namespace PitLeague.SimHub.Adapters.F1_25.Udp
             var buf = buffers[carIdx];
             buf.SetNumLaps(numLaps);
 
-            int lapCount = Math.Min(numLaps, MAX_LAPS);
+            int lapCount = Math.Min((int)numLaps, MAX_LAPS);
 
             for (int i = 0; i < lapCount; i++)
             {
